Show user

+++ app/Http/Controllers/SubjectController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Subject;
+use App\Schedule;
+use App\Group;
 
 class SubjectController extends Controller
 {
@@ -42,8 +44,16 @@ class SubjectController extends Controller
     public function delete($id) {
 
         $subject = Subject::find($id);
-        $subject->delete();
 
+        $sch = Schedule::where([['subject_id', $id]])->get();
+        $group = Group::where('subjects_id', 'LIKE', '%<'.$id.'>%')->get();
+
+        
+        if(count($sch) != 0 || count($group) != 0) {
+            return response()->json(false);    
+        }
+
+        $subject->delete();
         return response()->json(true);
     }
 }
